requirements.txt
shodan==1.29.1
python-nmap==0.7.1
numpy==1.24.2
